+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.2] - 2023-06-09
+
+### Added
+
+- An option to reload the editor window after creating a theme.
+- Template theme for making an easier development experience.
+
+### Changed
+
+- Icon for the extension.
+- Namespace, module and package syntax highlighting to fit surrounding colors.
+- Inactive bracket guide color to fit surrounding colors.
+
+### Fixed
+
+- Editor theme not updating when created from the command `codemos.createTheme`.
+- Wrong color attribute name `debugTokenExpression.pri` to `debugTokenExpression.value`.
+
 ## [1.0.1] - 2023-06-09
 
 ### Changed
@@ -22,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Initial release
 
-[unreleased]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/Codemos-Inc/codemos-modern/releases/tag/v1.0.0